requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
python-dotenv==1.2.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
pytz==2024.1

